jarvis/requirements.txt
python-telegram-bot>=20.0
google-generativeai>=0.3.0
requests
pytest
black
ruff
watchdog